managed to get mail to view working
