+++ app/templates/index.html
@@ -4,7 +4,7 @@
   <h1 class="text-center">Felist</h1>
 </div>
 <form action="" method="post">
-	<input type="submit" name="contact" value="Contact Us">
+	<input class="btn btn-primary" type="submit" name="contact" value="Contact Us">
 </form>
 
 {% endblock %}
